requirements.txt
appnope==0.1.0
decorator==4.1.2
Django==1.11.3
django-autofixture==0.12.1
django-ckeditor==5.3.0
django-extensions==1.8.1
django-js-asset==0.1.1
django-model-utils==2.5
django-widget-tweaks==1.4.1
factory-boy==2.7.0
fake-factory==9999.9.9
Faker==0.7.7
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.11.0
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.15
#psycopg2==2.6.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytils==0.3
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.3
traitlets==4.3.2
wcwidth==0.1.7
wheel==0.24.0
